feat: create a loaded replacement list widget that can be shared/used as needed anywhere fix: a bug if you attempted to use `get_replacements` method that could result in an error if one of the columns was empty
